requirements.txt
pyTelegramBotAPI
pandas
matplotlib
tabula-py
